requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
decorator==5.1.0
Django==3.2.7
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-location-field==2.1.0
django-taggit==2.0.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.7
